requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
ecs-logging==2.1.0
elastic-apm==6.18.0
elasticsearch==7.17.9
fastapi==0.103.1
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
h11==0.14.0
httptools==0.6.0
huggingface-hub
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.25.2
openai==0.28.0
packaging==23.1
Pillow==10.0.0
psutil==5.9.5
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
schematics==2.1.1
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.5.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.7.1
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
pytest==8.3.2
pytest-xdist==3.6.1
pytest-ordering==0.6
httpx==0.26.0
wheel
optimum
onnx
onnxruntime